Only inherit from controller on frontend controllers
