+++ app/controllers/spina/sitemaps_controller.rb
@@ -1,8 +1,8 @@
-module Spina
-  class SitemapsController < Spina::ApplicationController
-    def show
-      I18n.locale = I18n.default_locale
-      @pages = Page.live.sorted
-    end
+class Spina::SitemapsController < Spina::ApplicationController
+  
+  def show
+    I18n.locale = I18n.default_locale
+    @pages = Spina::Page.live.sorted
   end
+  
 end
